+++ wagtail_dripdrop/admin.py
@@ -1,1 +1,0 @@
-# Admin configuration for wagtail-dripdrop will be defined here.