+++ wagtail_dripdrop/models.py
@@ -1,1 +1,0 @@
-# Models for wagtail-dripdrop will be defined here.